Mixed guitar and arranged for loop
Tom mixing and arranging a bit to set up for a loop. Took some of the low-end out of the guitar tracks. Rolled the highs off the bass midi to get rid of the weird clip it would do on some notes. Looped the whole thing from the guitar part. Used the harp from the first empty section of the song cuz it was more the line that was played. Removed the harp from the 3 guitar section cuz it got too messy with it.